+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added filters based on annotations existence.
 - Nav link to the FTP site with seed MSA files and the MGnifams HMM library.
 - Home intro paragraph describing the MGnifams pipeline input data from the latest release of MGnify proteins (2024_04).