+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [v2.1.0dev] - unreleased
+## [v2.1.0] - 2026-04-28
 
 ### Added